requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
crewai==0.80.0
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
anthropic==0.40.0
